+++ Assets/PatchKit Patcher/Scripts/AppUpdater/IAppUpdaterStrategy.cs
@@ -2,8 +2,18 @@
 
 namespace PatchKit.Unity.Patcher.AppUpdater
 {
+    public enum StrategyType
+    {
+        None,
+        Empty,
+        Content,
+        Diff
+    }
+
     public interface IAppUpdaterStrategy
     {
         void Update(CancellationToken cancellationToken);
+
+        StrategyType GetStrategyType();
     }
 }