+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Added
 - HTTPS support for keys server
+- Fallback for content strategy in case of diff strategy failure
 
 ### Fixed
 - Compilation error on Unity 5.6 or higher